refactor: Change wrong data type in SQL script

+++ src/main/resources/create_table_account.sql
@@ -3,7 +3,7 @@
 CREATE TABLE IF NOT EXISTS account (
     account_id SERIAL PRIMARY KEY,
     account_name VARCHAR(255) NOT NULL,
-    balance DOUBLE NOT NULL,
+    balance decimal NOT NULL,
     currency_id INT NOT NULL,
     account_type "account_type" NOT NULL,
     FOREIGN KEY (currency_id) REFERENCES currency(currency_id)
